test for dev3_*
